producer/requirements.txt
kafka-python
prometheus_client
requests
sseclient-py